Wrote extract_categories and generated a categories database.
This new script extracts what suffix and subdomain MercadoLivre requires
for a search to be limited to a specific category. It builds a list of
parent categories, each of which has sublists of categories, which con-
tain the necessary information for them to be used in the searches. The
results are put in the pickled file "categories.pickle". This file will
be our categories database. It shouldn't change too often, so the script
doesn't have to be run again for a while.