Minor changes to ML_extract_categories, major changes to ML_scraper.py
Fixed the issue of categories having same ID, and then generated a new
categories database with slightly different structure: now the parent
category has a number, and it's children are identified by the parent's
number and their own number inside that parent category, separated by a
dot.

In the file ML_scraper, made the first implementation of the use of the
categories database, and encapsulated the extraction of the pages of re-
sults.

Also introduced first version of the input of the search parame-
ters. The input text is in portuguese, as the website in which this
script operates is in portuguese and wouldn't make sense to make english
menus for it.

The script still needs to be tested.